Extract assertion as test helper
Since it's shared in at least two different test modules.

See #2147
See #2337
