+++ test/support/test_helpers.ex
@@ -11,7 +11,7 @@ defmodule Ask.TestHelpers do
         ChannelBrokerAgent
       }
 
-      alias Ask.{PanelSurvey, Repo, Respondent, Survey, TestChannel}
+      alias Ask.{PanelSurvey, Repo, Respondent, Survey, SurveyLogEntry, TestChannel}
 
       @foo_string "foo"
       @bar_string "bar"
@@ -150,6 +150,16 @@ defmodule Ask.TestHelpers do
         assert a == active
         assert p == pending
       end
+
+      def assert_disposition_changed(respondent_id, old_disposition, new_disposition) do
+        last_entry =
+          Repo.one(from log in SurveyLogEntry, where: log.respondent_id == ^respondent_id, where: log.action_type == "disposition changed", order_by: [desc: :id], limit: 1)
+
+        assert last_entry.disposition == to_string(old_disposition)
+        assert last_entry.action_data == upcaseFirst(new_disposition)
+      end
+      defp upcaseFirst(value) when is_atom(value), do: to_string(value) |> upcaseFirst
+      defp upcaseFirst(<<first::utf8, rest::binary>>), do: String.upcase(<<first::utf8>>) <> rest
 
       defp broker_poll(), do: SurveyBroker.handle_info(:poll, nil)
